+++ src/include/boxmodus.h
@@ -50,7 +50,7 @@ class BoxModus : public ModusDefault {
 
   /** creates initial conditions from the particles.
    */
-  float initial_conditions(Particles *particles,
+  double initial_conditions(Particles *particles,
                           const ExperimentParameters &parameters);
 
   /** Enforces that all particles are inside the box
@@ -68,7 +68,7 @@ class BoxModus : public ModusDefault {
 
   /// \copydoc Smash::ModusDefault::create_grid
   Grid<GridOptions::PeriodicBoundaries> create_grid(
-      const Particles &particles, float min_cell_length,
+      const Particles &particles, double min_cell_length,
       CellSizeStrategy strategy = CellSizeStrategy::Optimal) const {
     return {{{0, 0, 0}, {length_, length_, length_}},
             particles,
@@ -79,35 +79,35 @@ class BoxModus : public ModusDefault {
   /// \copydoc Smash::ModusDefault::create_grandcan_thermalizer
   std::unique_ptr<GrandCanThermalizer> create_grandcan_thermalizer(
                                                Configuration& conf) const {
-    const std::array<float, 3> lat_size = {length_, length_, length_};
-    const std::array<float, 3> origin = {0.0f, 0.0f, 0.0f};
+    const std::array<double, 3> lat_size = {length_, length_, length_};
+    const std::array<double, 3> origin = {0.0f, 0.0f, 0.0f};
     const bool periodicity = true;
     return make_unique<GrandCanThermalizer>(conf, lat_size, origin,
         periodicity);
   }
 
   /// \copydoc Smash::ModusDefault::max_timestep()
-  float max_timestep(float max_transverse_distance_sqr) const {
+  double max_timestep(double max_transverse_distance_sqr) const {
     return 0.5f*std::sqrt(length_*length_ - max_transverse_distance_sqr);
   }
 
-  float length() const { return length_; }
+  double length() const { return length_; }
 
  private:
   const BoxInitialCondition initial_condition_;
   /// length of the cube's edge in fm/c
-  const float length_;
+  const double length_;
   /// Temperature of the Box in GeV
-  const float temperature_;
+  const double temperature_;
   /// initial time of the box
-  const float start_time_ = 0.0f;
+  const double start_time_ = 0.0f;
   /** whether to use a thermal initialization for all particles
    *  instead of specific numbers */
   const bool use_thermal_ = false;
   /// baryon chemical potential for thermal box
-  const float mub_;
+  const double mub_;
   /// strange chemical potential for thermal box
-  const float mus_;
+  const double mus_;
   /// particle multiplicities at initialization
   const std::map<PdgCode, int> init_multipl_;
